New approach to remove contamination ASVs
The new approach is based on a file with control sample information (assuming 3 possible control groups).
The central function for this is remove_contamination(), which is big and a bit complex, thus I prefer to call it using source instead of showing it in the main text (also so as to avoid modification by the user).